requirements.txt
streamlit
     langchain
     langchain-community
     langchain-ollama
     sentence-transformers
     faiss-cpu
     pdfplumber
     python-dotenv